+++ packages/babel-preset-next/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.12](https://github.com/easyops-cn/next-core/compare/@next-core/babel-preset-next@1.0.11...@next-core/babel-preset-next@1.0.12) (2024-01-02)
+
+
+### Bug Fixes
+
+* add missing dep `@babel/runtime` ([436c623](https://github.com/easyops-cn/next-core/commit/436c62386335b46194e93a823a26dfbbf67828f2))
+
+
+
+
+
+## [1.0.11](https://github.com/easyops-cn/next-core/compare/@next-core/babel-preset-next@1.0.10...@next-core/babel-preset-next@1.0.11) (2024-01-02)
+
+**Note:** Version bump only for package @next-core/babel-preset-next
+
+
+
+
+
 ## [1.0.10](https://github.com/easyops-cn/next-core/compare/@next-core/babel-preset-next@1.0.9...@next-core/babel-preset-next@1.0.10) (2023-12-25)
 
 **Note:** Version bump only for package @next-core/babel-preset-next