+++ packages/http/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.8](https://github.com/easyops-cn/next-core/compare/@next-core/http@1.1.7...@next-core/http@1.1.8) (2024-01-02)
+
+**Note:** Version bump only for package @next-core/http
+
+
+
+
+
 ## [1.1.7](https://github.com/easyops-cn/next-core/compare/@next-core/http@1.1.6...@next-core/http@1.1.7) (2023-12-25)
 
 **Note:** Version bump only for package @next-core/http