+++ packages/inject/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.27](https://github.com/easyops-cn/next-core/compare/@next-core/inject@1.0.26...@next-core/inject@1.0.27) (2024-01-10)
+
+**Note:** Version bump only for package @next-core/inject
+
+
+
+
+
+## [1.0.26](https://github.com/easyops-cn/next-core/compare/@next-core/inject@1.0.25...@next-core/inject@1.0.26) (2024-01-02)
+
+**Note:** Version bump only for package @next-core/inject
+
+
+
+
+
 ## [1.0.25](https://github.com/easyops-cn/next-core/compare/@next-core/inject@1.0.24...@next-core/inject@1.0.25) (2023-12-25)
 
 **Note:** Version bump only for package @next-core/inject