+++ packages/pipes/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.15](https://github.com/easyops-cn/next-core/compare/@next-core/pipes@2.0.14...@next-core/pipes@2.0.15) (2024-01-02)
+
+**Note:** Version bump only for package @next-core/pipes
+
+
+
+
+
 ## [2.0.14](https://github.com/easyops-cn/next-core/compare/@next-core/pipes@2.0.13...@next-core/pipes@2.0.14) (2023-12-25)
 
 **Note:** Version bump only for package @next-core/pipes