+++ packages/utils/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.6.9](https://github.com/easyops-cn/next-core/compare/@next-core/utils@1.6.8...@next-core/utils@1.6.9) (2024-01-10)
+
+**Note:** Version bump only for package @next-core/utils
+
+
+
+
+
+## [1.6.8](https://github.com/easyops-cn/next-core/compare/@next-core/utils@1.6.7...@next-core/utils@1.6.8) (2024-01-02)
+
+**Note:** Version bump only for package @next-core/utils
+
+
+
+
+
 ## [1.6.7](https://github.com/easyops-cn/next-core/compare/@next-core/utils@1.6.6...@next-core/utils@1.6.7) (2023-12-25)
 
 **Note:** Version bump only for package @next-core/utils