+++ packages/yo/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.37](https://github.com/easyops-cn/next-core/compare/@next-core/yo@1.5.36...@next-core/yo@1.5.37) (2024-01-31)
+
+**Note:** Version bump only for package @next-core/yo
+
+
+
+
+
+## [1.5.36](https://github.com/easyops-cn/next-core/compare/@next-core/yo@1.5.35...@next-core/yo@1.5.36) (2024-01-10)
+
+**Note:** Version bump only for package @next-core/yo
+
+
+
+
+
+## [1.5.35](https://github.com/easyops-cn/next-core/compare/@next-core/yo@1.5.34...@next-core/yo@1.5.35) (2024-01-04)
+
+**Note:** Version bump only for package @next-core/yo
+
+
+
+
+
+## [1.5.34](https://github.com/easyops-cn/next-core/compare/@next-core/yo@1.5.33...@next-core/yo@1.5.34) (2024-01-03)
+
+**Note:** Version bump only for package @next-core/yo
+
+
+
+
+
+## [1.5.33](https://github.com/easyops-cn/next-core/compare/@next-core/yo@1.5.32...@next-core/yo@1.5.33) (2024-01-02)
+
+**Note:** Version bump only for package @next-core/yo
+
+
+
+
+
 ## [1.5.32](https://github.com/easyops-cn/next-core/compare/@next-core/yo@1.5.31...@next-core/yo@1.5.32) (2023-12-28)
 
 **Note:** Version bump only for package @next-core/yo